move constants to sav
