+++ consts/consts.go
@@ -23,8 +23,3 @@ const (
 	BATTLE_STATS_LEVEL = 0x4
 	BATTLE_STATS_STAT = 0x8
 )
-
-const (
-	MAGIC_TIMESTAMP_JP_INTL = 0x20060623
-	MAGIC_TIMESTAMP_KR = 0x20070903
-)
